remove serial_test dependency
