+++ opentelemetry-otlp/tests/integration_test/Cargo.toml
@@ -17,7 +17,6 @@ ctor = "0.2.9"
 uuid = { version = "1.3", features = ["v4"] }
 tracing-subscriber = { workspace = true, features = ["env-filter","registry", "std", "fmt"] }
 tracing = {workspace = true}
-serial_test = "3.2.0"
 
 [target.'cfg(unix)'.dependencies]
 opentelemetry-appender-tracing = { path = "../../../opentelemetry-appender-tracing", default-features = false}
