+++ opentelemetry-otlp/tests/integration_test/tests/logs.rs
@@ -89,42 +89,18 @@ mod logtests {
 
     #[tokio::test(flavor = "multi_thread", worker_threads = 4)]
     #[cfg(any(feature = "tonic-client", feature = "reqwest-blocking-client"))]
-    #[serial_test::serial]
-    pub async fn test_logs() -> Result<()> {
-        // Make sure the container is running
-        use crate::{assert_logs_results, init_logs};
-        test_utils::start_collector_container().await?;
-        test_utils::cleanup_file("./actual/logs.json"); // Ensure logs.json is empty before the test
-        let logger_provider = init_logs(false).unwrap();
-        let layer = layer::OpenTelemetryTracingBridge::new(&logger_provider);
-        let subscriber = tracing_subscriber::registry().with(layer);
-        {
-            let _guard = tracing::subscriber::set_default(subscriber);
-            info!(target: "my-target", "hello from {}. My price is {}.", "banana", 2.99);
-        }
-        let _ = logger_provider.shutdown();
-        tokio::time::sleep(Duration::from_secs(5)).await;
-        assert_logs_results(test_utils::LOGS_FILE, "expected/logs.json")?;
-        Ok(())
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-    #[cfg(any(feature = "tonic-client", feature = "reqwest-blocking-client"))]
-    #[serial_test::parallel]
     pub async fn logs_batch_tokio_multi_thread() -> Result<()> {
         logs_tokio_helper(false).await
     }
 
     #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
     #[cfg(any(feature = "tonic-client", feature = "reqwest-blocking-client"))]
-    #[serial_test::parallel]
     pub async fn logs_batch_tokio_multi_with_one_worker() -> Result<()> {
         logs_tokio_helper(false).await
     }
 
     #[tokio::test(flavor = "current_thread")]
     #[cfg(any(feature = "tonic-client", feature = "reqwest-blocking-client"))]
-    #[serial_test::parallel]
     pub async fn logs_batch_tokio_current() -> Result<()> {
         logs_tokio_helper(false).await
     }
@@ -151,14 +127,12 @@ mod logtests {
 
     #[tokio::test(flavor = "multi_thread", worker_threads = 4)]
     #[cfg(any(feature = "tonic-client", feature = "reqwest-client"))]
-    #[serial_test::parallel]
     pub async fn logs_simple_tokio_multi_thread() -> Result<()> {
         logs_tokio_helper(true).await
     }
 
     #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
     #[cfg(any(feature = "tonic-client", feature = "reqwest-client"))]
-    #[serial_test::parallel]
     pub async fn logs_simple_tokio_multi_with_one_worker() -> Result<()> {
         logs_tokio_helper(true).await
     }
@@ -167,14 +141,12 @@ mod logtests {
     #[ignore]
     #[tokio::test(flavor = "current_thread")]
     #[cfg(any(feature = "tonic-client", feature = "reqwest-client"))]
-    #[serial_test::parallel]
     pub async fn logs_simple_tokio_current() -> Result<()> {
         logs_tokio_helper(true).await
     }
 
     #[test]
     #[cfg(any(feature = "tonic-client", feature = "reqwest-blocking-client"))]
-    #[serial_test::parallel]
     pub fn logs_batch_non_tokio_main() -> Result<()> {
         logs_non_tokio_helper(false)
     }
@@ -207,7 +179,6 @@ mod logtests {
 
     #[test]
     #[cfg(any(feature = "tonic-client", feature = "reqwest-blocking-client"))]
-    #[serial_test::parallel]
     pub fn logs_simple_non_tokio_main() -> Result<()> {
         logs_non_tokio_helper(true)
     }
